Moving Sizing & Timing classes from Server to Configuration
