+++ src/test/java/io/vlingo/http/resource/ServerTest.java
@@ -19,8 +19,8 @@ import io.vlingo.actors.Definition;
 import io.vlingo.actors.testkit.TestUntil;
 import io.vlingo.http.Response;
 import io.vlingo.http.ResponseHeader;
-import io.vlingo.http.resource.Server.Sizing;
-import io.vlingo.http.resource.Server.Timing;
+import io.vlingo.http.resource.Configuration.Sizing;
+import io.vlingo.http.resource.Configuration.Timing;
 import io.vlingo.http.resource.TestResponseChannelConsumer.Progress;
 import io.vlingo.http.sample.user.model.User;
 import io.vlingo.wire.channel.ResponseChannelConsumer;
